requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
sendgrid==6.11.0
reportlab==4.2.2
pdfplumber==0.11.4
phonenumbers>=8.13.0
google-cloud-pubsub>=2.21.0
twilio>=8.0.0
pytest>=7.4.0
pytest-flask>=1.3.0
